+++ ditto/lastfm/views.py
@@ -1,7 +1,7 @@
 from datetime import timedelta
 
 from django.http import Http404
-from django.utils.translation import ugettext as _
+from django.utils.translation import gettext as _
 from django.views.generic import DetailView, TemplateView
 from django.views.generic.detail import SingleObjectMixin