+++ ditto/twitter/management/commands/import_twitter_tweets.py
@@ -3,7 +3,7 @@ import os
 
 from django.core.management.base import BaseCommand, CommandError
 
-from ...ingest import TweetIngester
+from ...ingest import Version1TweetIngester, Version2TweetIngester
 
 
 class Command(BaseCommand):
@@ -24,28 +24,52 @@ class Command(BaseCommand):
             help="Path to the directory that is the archive",
         )
 
+        parser.add_argument(
+            "--archive-version",
+            action="store",
+            default=None,
+            help="v1 or v2 (default). Which format of archives to import from.",
+        )
+
     def handle(self, *args, **options):
         # Location of the directory holding the tweet JSON files within the
         # archive:
-        subpath = "/data/js/tweets"
+
+        ingester_class = None
+
+        # For v2, the default:
+        # Where the JS files are:
+        subpath = "/data"
+        ingester_class = Version2TweetIngester
+
+        if options["archive_version"]:
+            if options["archive_version"] == "v1":
+                # Where the JS files are:
+                subpath = "/data/js/tweets"
+                ingester_class = Version1TweetIngester
+
+            elif options["archive_version"] != "v2":
+                raise CommandError(
+                    f"version should be v1 or v2, not '{options['archive_version']}"
+                )
 
         if options["path"]:
             if os.path.isdir(options["path"]):
-                tweets_dir = "%s%s" % (options["path"], subpath)
-                if os.path.isdir(tweets_dir):
-                    result = TweetIngester().ingest(directory=tweets_dir)
+                js_dir = "%s%s" % (options["path"], subpath)
+                if os.path.isdir(js_dir):
+                    result = ingester_class().ingest(directory=js_dir)
                 else:
                     raise CommandError(
-                        "Expected to find a directory at '%s' containing JSON files"
-                        % tweets_dir
+                        f"Expected to find a directory at '{js_dir}' "
+                        "containing .js file(s)"
                     )
             else:
-                raise CommandError("Can't find a directory at '%s'" % options["path"])
+                raise CommandError(f"Can't find a directory at '{options['path']}'")
         else:
             raise CommandError(
                 (
-                    "Specify the location of the archive, "
-                    "e.g. --path=/Path/To/1234567890_abcdefg12345"
+                    "Specify the location of the archive directory, "
+                    "e.g. --path=/path/to/twitter-2022-01-31-abcdef123456"
                 )
             )
 
@@ -53,13 +77,13 @@ class Command(BaseCommand):
             if result["success"]:
                 tweetnoun = "tweet" if result["tweets"] == 1 else "tweets"
                 filenoun = "file" if result["files"] == 1 else "files"
+                mediafilenoun = "file" if result["media"] == 1 else "files"
 
                 self.stdout.write(
-                    "Imported %s %s from %s %s"
-                    % (result["tweets"], tweetnoun, result["files"], filenoun)
+                    f"Imported {result['tweets']} {tweetnoun} from "
+                    f"{result['files']} {filenoun}, "
+                    f"and {result['media']} media {mediafilenoun}"
                 )
             else:
 
-                self.stderr.write(
-                    "Failed to import tweets: %s" % (result["messages"][0])
-                )
+                self.stderr.write(f"Failed to import tweets: {result['messages'][0]}")