+++ tests/pinboard/test_templatetags.py
@@ -127,9 +127,9 @@ class PopularBookmarkTagsTestCase(TestCase):
     def test_tags(self):
         "Contains the correct data"
         bookmark_1 = BookmarkFactory()
-        bookmark_1.tags.set("fish", "carp")
+        bookmark_1.tags.set(["fish", "carp"])
         bookmark_2 = BookmarkFactory()
-        bookmark_2.tags.set("fish", "cod")
+        bookmark_2.tags.set(["fish", "cod"])
         tags = ditto_pinboard.popular_bookmark_tags()
         self.assertEqual(len(tags), 3)
         self.assertEqual(tags[0].name, "fish")
@@ -142,9 +142,9 @@ class PopularBookmarkTagsTestCase(TestCase):
     def test_tags_privacy_bookmarks(self):
         "Doesn't display tags from private bookmarks"
         bookmark_1 = BookmarkFactory(is_private=True)
-        bookmark_1.tags.set("fish", "carp")
+        bookmark_1.tags.set(["fish", "carp"])
         bookmark_2 = BookmarkFactory(is_private=False)
-        bookmark_2.tags.set("fish", "cod")
+        bookmark_2.tags.set(["fish", "cod"])
         tags = ditto_pinboard.popular_bookmark_tags()
         self.assertEqual(len(tags), 2)
         self.assertEqual(tags[0].name, "fish")
@@ -155,7 +155,7 @@ class PopularBookmarkTagsTestCase(TestCase):
     def test_tags_privacy_tags(self):
         "Doesn't display private .tags"
         bookmark = BookmarkFactory()
-        bookmark.tags.set("ispublic", ".notpublic", "alsopublic")
+        bookmark.tags.set(["ispublic", ".notpublic", "alsopublic"])
         tags = ditto_pinboard.popular_bookmark_tags()
         self.assertEqual(len(tags), 2)
         # Tags are ordered by popularity, so can't be sure
@@ -167,13 +167,13 @@ class PopularBookmarkTagsTestCase(TestCase):
     def test_tags_limit_default(self):
         "It should return 10 tags by default"
         bookmark = BookmarkFactory()
-        bookmark.tags.set("1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "11")
+        bookmark.tags.set(["1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "11"])
         tags = ditto_pinboard.popular_bookmark_tags()
         self.assertEqual(len(tags), 10)
 
     def test_tags_limit_custom(self):
         "It should return `limit` tags"
         bookmark = BookmarkFactory()
-        bookmark.tags.set("1", "2", "3", "4", "5")
+        bookmark.tags.set(["1", "2", "3", "4", "5"])
         tags = ditto_pinboard.popular_bookmark_tags(limit=3)
         self.assertEqual(len(tags), 3)