+++ tests/twitter/fixtures/ingest/v2_no_tweets/account.js
@@ -1,0 +1,12 @@
+window.YTD.account.part0 = [
+  {
+    "account" : {
+      "email" : "phil@gyford.com",
+      "createdVia" : "web",
+      "username" : "philgyford",
+      "accountId" : "12552",
+      "createdAt" : "2006-11-15T16:55:59.000Z",
+      "accountDisplayName" : "Phil Gyford"
+    }
+  }
+]