+++ tests/twitter/fixtures/ingest/v2_no_tweets/tweet.js
@@ -1,0 +1,2 @@
+window.YTD.tweet.part0 = [
+]