+++ tests/twitter/fixtures/ingest/v2_with_media/profile.js
@@ -1,0 +1,13 @@
+window.YTD.profile.part0 = [
+  {
+    "profile" : {
+      "description" : {
+        "bio" : "Creator of the Bishop of Manchester’s favourite meme // Also @samuelpepys and @todaysguardian",
+        "website" : "https://t.co/FsYzXrATit",
+        "location" : "Herefordshire, UK"
+      },
+      "avatarMediaUrl" : "https://pbs.twimg.com/profile_images/1167616130/james_200208_300x300.jpg",
+      "headerMediaUrl" : "https://pbs.twimg.com/profile_banners/12552/1603038696"
+    }
+  }
+]