+++ tests/twitter/fixtures/ingest/v2_with_media/tweet.js
@@ -1,0 +1,243 @@
+window.YTD.tweet.part0 = [
+  {
+     "tweet" : {
+       "retweeted" : false,
+       "source" : "<a href=\"https://tapbots.com/software/tweetbot/mac\" rel=\"nofollow\">Tweetbot for Mac</a>",
+       "entities" : {
+         "user_mentions" : [ ],
+         "urls" : [
+           {
+             "url" : "https://t.co/piSTKg40ct",
+             "expanded_url" : "http://Dropbox.com",
+             "display_url" : "Dropbox.com",
+             "indices" : [
+               "109",
+               "132"
+             ]
+           }
+         ],
+         "symbols" : [ ],
+         "media" : [
+           {
+             "expanded_url" : "https://twitter.com/philgyford/status/1247471193357275137/photo/1",
+             "indices" : [
+               "237",
+               "260"
+             ],
+             "url" : "https://t.co/DcKCkNekxp",
+             "media_url" : "http://pbs.twimg.com/tweet_video_thumb/EU_oaKjWkAAj6Gv.jpg",
+             "id_str" : "1247471158011793408",
+             "id" : "1247471158011793408",
+             "media_url_https" : "https://pbs.twimg.com/tweet_video_thumb/EU_oaKjWkAAj6Gv.jpg",
+             "sizes" : {
+               "small" : {
+                 "w" : "250",
+                 "h" : "330",
+                 "resize" : "fit"
+               },
+               "thumb" : {
+                 "w" : "150",
+                 "h" : "150",
+                 "resize" : "crop"
+               },
+               "medium" : {
+                 "w" : "250",
+                 "h" : "330",
+                 "resize" : "fit"
+               },
+               "large" : {
+                 "w" : "250",
+                 "h" : "330",
+                 "resize" : "fit"
+               }
+             },
+             "type" : "photo",
+             "display_url" : "pic.twitter.com/DcKCkNekxp"
+           }
+         ],
+         "hashtags" : [ ]
+       },
+       "display_text_range" : [
+         "0",
+         "260"
+       ],
+       "favorite_count" : "1",
+       "id_str" : "1247471193357275137",
+       "truncated" : false,
+       "retweet_count" : "0",
+       "id" : "1247471193357275137",
+       "possibly_sensitive" : false,
+       "created_at" : "Tue Apr 07 10:28:04 +0000 2020",
+       "favorited" : false,
+       "full_text" : "The good thing is that once I've finished the laborious process of downloading 39 separate shared folders on https://t.co/piSTKg40ct, the day can           only get better.\n\nThe weirdly-changing top-level navigation is the least confusing aspect. https://t.co/DcKCkNekxp",
+       "lang" : "en",
+       "extended_entities" : {
+         "media" : [
+           {
+             "expanded_url" : "https://twitter.com/philgyford/status/1247471193357275137/photo/1",
+             "indices" : [
+               "237",
+               "260"
+             ],
+             "url" : "https://t.co/DcKCkNekxp",
+             "media_url" : "http://pbs.twimg.com/tweet_video_thumb/EU_oaKjWkAAj6Gv.jpg",
+             "id_str" : "1247471158011793408",
+             "video_info" : {
+               "aspect_ratio" : [
+                 "25",
+                 "33"
+               ],
+               "variants" : [
+                 {
+                   "bitrate" : "0",
+                   "content_type" : "video/mp4",
+                   "url" : "https://video.twimg.com/tweet_video/EU_oaKjWkAAj6Gv.mp4"
+                 }
+               ]
+             },
+             "id" : "1247471158011793408",
+             "media_url_https" : "https://pbs.twimg.com/tweet_video_thumb/EU_oaKjWkAAj6Gv.jpg",
+             "sizes" : {
+               "small" : {
+                 "w" : "250",
+                 "h" : "330",
+                 "resize" : "fit"
+               },
+               "thumb" : {
+                 "w" : "150",
+                 "h" : "150",
+                 "resize" : "crop"
+               },
+               "medium" : {
+                 "w" : "250",
+                 "h" : "330",
+                 "resize" : "fit"
+               },
+               "large" : {
+                 "w" : "250",
+                 "h" : "330",
+                 "resize" : "fit"
+               }
+             },
+             "type" : "animated_gif",
+             "display_url" : "pic.twitter.com/DcKCkNekxp"
+           }
+         ]
+       }
+     }
+   },
+   {
+     "tweet" : {
+       "retweeted" : false,
+       "source" : "<a href=\"http://twitter.com/#!/download/ipad\" rel=\"nofollow\">Twitter for iPad</a>",
+       "entities" : {
+         "user_mentions" : [ ],
+         "urls" : [
+           {
+             "url" : "https://t.co/6hPhxAcr9n",
+             "expanded_url" : "https://twitter.com/peterme/status/1358811736464191488",
+             "display_url" : "twitter.com/peterme/status…",
+             "indices" : [
+               "179",
+               "202"
+             ]
+           }
+         ],
+         "symbols" : [ ],
+         "media" : [
+           {
+             "expanded_url" : "https://twitter.com/philgyford/status/1359135226161750021/photo/1",
+             "indices" : [
+               "203",
+               "226"
+             ],
+             "url" : "https://t.co/0tWLxyETsw",
+             "media_url" : "http://pbs.twimg.com/media/EtyeQ0FWYAEQDqk.jpg",
+             "id_str" : "1359135199255224321",
+             "id" : "1359135199255224321",
+             "media_url_https" : "https://pbs.twimg.com/media/EtyeQ0FWYAEQDqk.jpg",
+             "sizes" : {
+               "large" : {
+                 "w" : "248",
+                 "h" : "256",
+                 "resize" : "fit"
+               },
+               "thumb" : {
+                 "w" : "150",
+                 "h" : "150",
+                 "resize" : "crop"
+               },
+               "medium" : {
+                 "w" : "248",
+                 "h" : "256",
+                 "resize" : "fit"
+               },
+               "small" : {
+                 "w" : "248",
+                 "h" : "256",
+                 "resize" : "fit"
+               }
+             },
+             "type" : "photo",
+             "display_url" : "pic.twitter.com/0tWLxyETsw"
+           }
+         ],
+         "hashtags" : [ ]
+       },
+       "display_text_range" : [
+         "0",
+         "226"
+       ],
+       "favorite_count" : "3",
+       "id_str" : "1359135226161750021",
+       "truncated" : false,
+       "retweet_count" : "0",
+       "id" : "1359135226161750021",
+       "possibly_sensitive" : false,
+       "created_at" : "Tue Feb 09 13:41:04 +0000 2021",
+       "favorited" : false,
+       "full_text" : "I don’t know why I’d reveal my age when mentioning this anyway, but I think it was QS Defender, at my friend Simon’s. He had a 16K RAM Pack, if you         can imagine such extravagance. https://t.co/6hPhxAcr9n https://t.co/0tWLxyETsw",
+       "lang" : "en",
+       "extended_entities" : {
+         "media" : [
+           {
+             "expanded_url" : "https://twitter.com/philgyford/status/1359135226161750021/photo/1",
+             "indices" : [
+               "203",
+               "226"
+             ],
+             "url" : "https://t.co/0tWLxyETsw",
+             "media_url" : "http://pbs.twimg.com/media/EtyeQ0FWYAEQDqk.jpg",
+             "id_str" : "1359135199255224321",
+             "id" : "1359135199255224321",
+             "media_url_https" : "https://pbs.twimg.com/media/EtyeQ0FWYAEQDqk.jpg",
+             "sizes" : {
+               "large" : {
+                 "w" : "248",
+                 "h" : "256",
+                 "resize" : "fit"
+               },
+               "thumb" : {
+                 "w" : "150",
+                 "h" : "150",
+                 "resize" : "crop"
+               },
+               "medium" : {
+                 "w" : "248",
+                 "h" : "256",
+                 "resize" : "fit"
+               },
+               "small" : {
+                 "w" : "248",
+                 "h" : "256",
+                 "resize" : "fit"
+               }
+             },
+             "type" : "photo",
+             "display_url" : "pic.twitter.com/0tWLxyETsw"
+           }
+         ]
+       }
+     }
+   }
+]