+++ tests/twitter/test_utils.py
@@ -31,7 +31,7 @@ class HtmlifyDescriptionTestCase(HtmlifyTestCase):
                 '#testing</a> <a href="http://www.gyford.com" rel="external">'
                 'gyford.com</a> $IBM <a href="https://twitter.com/search?q=%23test">'
                 '#test</a> <a href="https://twitter.com/philgyford">@philgyford</a>'
-            )
+            ),
         )
 
 
@@ -53,7 +53,7 @@ class HtmlifyTweetEntitiesTestCase(HtmlifyTestCase):
             (
                 'and <a href="http://www.wired.com/2015/10/meet-walking-dead-hp-'
                 'cisco-dell-emc-ibm-oracle/" rel="external">wired.com/2015/10/'
-                'meet-w…</a>'
+                "meet-w…</a>"
             )
             in tweet_html
         )
@@ -84,7 +84,7 @@ class HtmlifyTweetEntitiesTestCase(HtmlifyTestCase):
         self.assertEqual(
             (
                 '<a href="https://twitter.com/genmon" rel="external">'
-                '@genmon</a> lovely, on my way.'
+                "@genmon</a> lovely, on my way."
             ),
             tweet_html,
         )
@@ -114,7 +114,7 @@ class HtmlifyTweetEntitiesTestCase(HtmlifyTestCase):
         tweet_html = htmlify_tweet(self.getJson(api_fixture))
         self.assertTrue(
             (
-                '@denisewilton</a> '
+                "@denisewilton</a> "
                 '<a href="https://twitter.com/search?q=%23LOVEWHATYOUDO" '
                 'rel="external">#LOVEWHATYOUDO</a> '
                 '<a href="https://twitter.com/search?q=%23DOWHATYOULOVE" '
@@ -129,13 +129,13 @@ class HtmlifyTweetEntitiesTestCase(HtmlifyTestCase):
         self.assertEqual(
             tweet_html,
             (
-                'Some symbols: '
+                "Some symbols: "
                 '<a href="https://twitter.com/search?q=%24AAPL" rel="external">'
-                '$AAPL</a> and '
+                "$AAPL</a> and "
                 '<a href="https://twitter.com/search?q=%24PEP" rel="external">'
-                '$PEP</a> and $ANOTHER and '
+                "$PEP</a> and $ANOTHER and "
                 '<a href="https://twitter.com/search?q=%24A" rel="external">$A</a>.'
-            )
+            ),
         )
 
 
@@ -160,6 +160,48 @@ class HtmlifyTweetTestCase(HtmlifyTestCase):
         self.assertEqual("Y", tweet_html[-1])
 
 
+class HtmlifyTweetStrsNotIntsTestCase(HtmlifyTestCase):
+    def test_handles_display_text_range_str(self):
+        """Cope correctly if display_text_range is strings, not ints.
+
+        Some tweet JSON have the display_text_range set as strings not ints,
+        e.g. ["0", "140"] rather than [0, 140]. Particularly when the
+        JSON has come from the downloaded Twitter archive.
+        It should be able to cope with that.
+        """
+        api_fixture = "tweet_with_display_text_range_str.json"
+        tweet_html = htmlify_tweet(self.getJson(api_fixture))
+        self.assertEqual(
+            tweet_html,
+            (
+                "Open iPad Slack, scroll to bottom. Close Slack. Open Slack; it’s "
+                "scrolled back up. Every time. Maddening. Not biggest problem in "
+                "world, BUT."
+            ),
+        )
+
+    def test_handles_entities_indicies_str(self):
+        """Cope correctly if entities' indicies are strings, not ints.
+
+        Some tweet JSON have the ["entities"][<kind>]["indices"] set as
+        strings not ints, e.g. ["0", "9"] rather than [0, 9].
+        Particularly when the JSON has come from the downloaded Twitter
+        archive.
+        It should be able to cope with that.
+        """
+        api_fixture = "tweet_with_entities_indices_str.json"
+        tweet_html = htmlify_tweet(self.getJson(api_fixture))
+        self.assertEqual(
+            tweet_html,
+            (
+                '<a href="https://twitter.com/terrycol" rel="external">@terrycol</a> '
+                '<a href="https://twitter.com/bobferris" rel="external">@bobferris</a> '
+                "I liked it and only thought some of it was a bit silly. But analysis "
+                "beyond that is probably beyond the scope of Twitter."
+            ),
+        )
+
+
 class HtmlifyTweetUrlsTestCase(HtmlifyTestCase):
     "Further tests for specific problems with URLs."
 
@@ -173,11 +215,11 @@ class HtmlifyTweetUrlsTestCase(HtmlifyTestCase):
         self.assertEqual(
             tweet_html,
             (
-                'Made a little Twitter thing: '
+                "Made a little Twitter thing: "
                 '<a href="http://www.gyford.com/phil/writing/2006/12/02/'
                 'quick_twitter.php">'
-                'http://www.gyford.com/phil/writing/2006/12/02/quick_twitter.php</a>'
-            )
+                "http://www.gyford.com/phil/writing/2006/12/02/quick_twitter.php</a>"
+            ),
         )
 
     def test_urls_with_no_media(self):