+++ CHANGELOG.md
@@ -4,7 +4,22 @@ Changelog (Django Ditto)
 Unreleased
 ----------
 
-- None.
+- None
+
+
+2.0.0 - 2022-02-14
+------------------
+
+- **Backwards incompatible:** Drop support for Django 2.2 and 3.1.
+
+- **Backwards incompatible:** Requires django-taggit >= v2.0.0
+  (It changed how `TaggableManager` sets tags: https://github.com/jazzband/django-taggit/blob/master/CHANGELOG.rst#200-2021-11-14 )
+
+- Add support for Django 4.0 (#223)
+
+- Add support for Python 3.10 (#225)
+
+- Add support for importing the "new" (2019-onwwards) format of downloadable Twitter archive (#229).
 
 
 1.4.2 - 2021-10-22